Separate the popup from the main component
