+++ src/containers/artist-recommendations/ArtistRecommendationsPopup.module.css
@@ -1,85 +1,13 @@
-.icon {
-  padding: 2px;
-  margin: -6px;
-  box-sizing: content-box;
-  position: relative;
-  transition: all 0.17s ease-in-out !important;
-}
-.closeButton {
-  width: 16px;
-  margin-right: 16px;
-}
-.closeButton:hover {
-  cursor: pointer;
+.popup {
+  width: 352px;
 }
 
-.icon.remove g path {
-  fill: var(--neutral-light-4);
-}
-.header {
-  display: flex;
-}
 .headerTitle {
   color: var(--secondary);
   font-weight: var(--font-bold);
   font-size: var(--font-l);
 }
-.popup {
-  width: 352px;
-}
-.popupContent {
-  padding: 8px;
-  font-size: var(--font-s);
-}
-.popupContent > * {
+
+.popupItem {
   margin: 16px;
-}
-.popupContent p {
-  line-height: 1.25;
-}
-
-.profilePictureList {
-  display: flex;
-}
-
-.profilePictureWrapper {
-  height: 52px;
-  width: 52px;
-  flex: 0 0 52px;
-  margin-right: -7px;
-}
-/* .profilePictureWrapper:hover {
-  cursor: pointer;
-  margin-top: -4px;
-  transition: margin-top 0.1s linear;
-} */
-
-.profilePicture {
-  box-sizing: border-box;
-  height: 52px;
-  width: 52px;
-  flex: 0 0 52px;
-  border: 1px solid var(--neutral-light-8);
-  border-radius: 50%;
-
-  background-size: cover;
-  background-position: center;
-  background-repeat: no-repeat;
-  background-color: var(--default-profile-picture-background);
-}
-
-.artistLink {
-  display: inline-block;
-  color: var(--secondary);
-  font-weight: var(--font-bold);
-  margin-bottom: 4px;
-  cursor: pointer;
-  line-height: normal;
-}
-.artistLink :global(.artistPopover) {
-  display: inline-block;
-}
-
-.verified > *:first-child {
-  margin-left: 5px;
 }
